refactor(run): split orchestrator.py (983 LOC) into orchestrator_pkg/ (Campaign B5)
Splits the historical 983-LOC monolith at
plugins/shipwright-run/scripts/lib/orchestrator.py into a per-concern
package, leaving a 41-LOC re-export shim at the original path so
existing callers — `from orchestrator import X`, `python
orchestrator.py CMD`, `mocker.patch("orchestrator.X")`, every SKILL.md
that subprocesses this exact path — stay green.

Package layout (`orchestrator_pkg/`):

- constants — schema, paths, allowlists (74 LOC)
- events — record_event.py wrappers (70 LOC)
- config_io — load/save/migrate (51 LOC)
- legacy_migration — drop compliance/security from old configs (189)
- config_factory — build_pipeline + create_config (158 LOC)
- compliance_runner — run_compliance_update subprocess wrapper (93)
- critical_gates — Phase-Quality W5/W6/W7 (88 LOC)
- step_planning — get_next_step + update_step (194 LOC)
- build_progress — get_build_progress (74 LOC)
- router — F2 lifecycle subcommand dispatcher (133 LOC)
- cli — argparse + main (141 LOC)
- __init__.py — public re-exports (72 LOC)

Every submodule is well under the 300-LOC source budget. Baseline
entry removed in this commit per the campaign cleanup-invariant.

Backward-compat:

- Shim binds the patched-via-orchestrator-namespace names
  (run_compliance_update, _COMPLIANCE_SCRIPT, _record_*,
  _collect_critical_gate_issues, _enforce_critical_gates_enabled,
  _read_latest_phase_quality_finding, _CRITICAL_GATE_CHECK_IDS,
  _dispatch_lifecycle) so historical `mocker.patch("orchestrator.X")`
  tests continue to intercept callers in step_planning, legacy_migration
  and compliance_runner via late `sys.modules["orchestrator"]` lookup.
- Shim re-exports the full public surface (build_pipeline,
  create_config, get_build_progress, get_next_step, is_v2_config,
  load_run_config, run_compliance_update, save_run_config,
  update_step, main, plus CONFIG_NAME / PIPELINE_STEPS /
  SCHEMA_VERSION constants).
- `shared/scripts/sync_check.py` updated to look for the PIPELINE_STEPS
  literal in `orchestrator_pkg/constants.py` first, with fallback to
  the legacy monolith path for branches that haven't picked up the
  split yet.
- `docs/hooks-and-pipeline.md` updated to point at the new locations.

Empirical verification:

- 35 / 35 `test_orchestrator.py` tests pass untouched (grandfathered
  file not modified per campaign rule).
- 7 new parity probes in `test_orchestrator_split_parity.py` (public
  names importable, patched-name binding, end-to-end mock intercept,
  router command set unchanged, fail-closed reasons unchanged, 2-phase
  flow routes identically, PIPELINE_STEPS literal in constants.py).
- 144 / 144 plugin tests + 24 / 24 orchestrator-touching integration
  tests + 64 / 64 shared phase-quality / hook / context tests pass.

External LLM review (mode=iterate + mode=code, OpenRouter): high-severity
findings on layout deviation (used `orchestrator_pkg/` not literal
`orchestrator/` to avoid file-vs-package shadowing), missing
`_dispatch_lifecycle` re-export, missing 2-phase routing integration
probe — all three addressed (alias added on shim, integration probe
added in parity suite, layout deviation documented in __init__.py).

Refs Campaign B5 — `iterate/bloat-B-orchestrator-split`
